requirements.txt
# requirements.txt
python-kasa>=0.10.2
python-telegram-bot==13.7
aiohttp>=3.11.0
nest_asyncio>=1.5.8